+++ tests/test_arcDataOps.cpp
@@ -502,7 +502,7 @@ bool pieceVecsAreEqual(std::vector<ArcPiece> v1, std::vector<ArcPiece> v2){
 	if(v1.size() != v2.size())
 		return false;
 
-	for(size_t i = 0; i < v1.size(); i++){
+	for(unsigned int i = 0; i < v1.size(); i++){
 		if(v1[i] != v2[i])
 			return false;
 	}
